Change project layout
