+++ clementine/config/__init__.py
@@ -1,0 +1,5 @@
+"""Configuration utilities for the Clementine bot."""
+
+from .logging import LoggingConfigurator
+
+__all__ = ["LoggingConfigurator"]
